refactor(dashboard): move vcs and realtime from apps/ to utils/
These modules are infrastructure utilities (webhook signature verification,
WebSocket broadcasting) that don't follow the Django-style app structure.
Moving them to utils/ better reflects their role in the codebase.
