+++ dashboard/src/apps.rs
@@ -10,8 +10,4 @@ pub mod clusters;
 #[cfg(native)]
 pub mod deployments;
 #[cfg(native)]
-pub mod realtime;
-#[cfg(native)]
 pub mod validators;
-#[cfg(native)]
-pub mod vcs;
