+++ dashboard/src/config/urls.rs
@@ -36,7 +36,7 @@ use reinhardt::{WebSocketRoute, WebSocketRouter, register_websocket_router};
 
 use crate::apps::auth::server;
 use crate::apps::auth::services::local_auth::LocalAuthService;
-use crate::apps::realtime::broadcaster::WsBroadcaster;
+use crate::utils::realtime::broadcaster::WsBroadcaster;
 use crate::config::middleware::CspPathMiddleware;
 use reinhardt::{
 	CookieSessionAuthMiddleware, CookieSessionConfig, OriginGuardMiddleware, RedisSessionBackend,
