+++ dashboard/src/utils.rs
@@ -1,0 +1,8 @@
+//! Utility modules for Reinhardt Cloud.
+//!
+//! Infrastructure utilities that don't follow the Django-style app structure.
+
+#[cfg(native)]
+pub mod realtime;
+#[cfg(native)]
+pub mod vcs;
