+++ dashboard/src/utils/realtime.rs
