+++ dashboard/src/utils/vcs.rs
@@ -1,0 +1,7 @@
+//! VCS (Version Control System) utility module.
+//!
+//! Provides webhook signature verification and event parsing
+//! for GitHub and GitLab integrations.
+
+pub mod events;
+pub mod signature;
